+++ src/main/java/svenhjol/charmony/core/base/Setup.java
@@ -1,0 +1,23 @@
+package svenhjol.charmony.core.base;
+
+
+public abstract class Setup<F extends SidedFeature> {
+    private final F feature;
+
+    public Setup(F feature) {
+        this.feature = feature;
+        this.feature.mod().addBootStep(feature, boot());
+    }
+
+    public F feature() {
+        return this.feature;
+    }
+
+    public Log log() {
+        return this.feature.log();
+    }
+
+    public Runnable boot() {
+        return () -> {};
+    }
+}